requirements.txt
python-dotenv
langgraph==0.1.16
langchain
langchain_core
langchain_openai
IPython
pyyaml
pydantic
streamlit
loguru
azure_devops
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
bluesheperd